ow09

+++ src/minecraft/net/minecraft/block/BlockAir.java
@@ -45,9 +45,6 @@ public class BlockAir extends Block
 
     /**
      * Spawns this Block's drops into the World as EntityItems.
-     *  
-     * @param chance The chance that each Item is actually spawned (1.0 = always, 0.0 = never)
-     * @param fortune The player's fortune level
      */
     public void dropBlockAsItemWithChance(World worldIn, BlockPos pos, IBlockState state, float chance, int fortune)
     {
@@ -65,7 +62,7 @@ public class BlockAir extends Block
     {
         if (!mapOriginalOpacity.containsKey(p_setLightOpacity_0_))
         {
-            mapOriginalOpacity.put(p_setLightOpacity_0_, Integer.valueOf(p_setLightOpacity_0_.lightOpacity));
+            mapOriginalOpacity.put(p_setLightOpacity_0_, p_setLightOpacity_0_.lightOpacity);
         }
 
         p_setLightOpacity_0_.lightOpacity = p_setLightOpacity_1_;
@@ -75,7 +72,7 @@ public class BlockAir extends Block
     {
         if (mapOriginalOpacity.containsKey(p_restoreLightOpacity_0_))
         {
-            int i = ((Integer)mapOriginalOpacity.get(p_restoreLightOpacity_0_)).intValue();
+            int i = (int) mapOriginalOpacity.get(p_restoreLightOpacity_0_);
             setLightOpacity(p_restoreLightOpacity_0_, i);
         }
     }
